server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
websockets>=11.0






'''openenv[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0



